+++ static/app/components/omniSearch/animation/lottieAnimation.tsx
