+++ static/app/components/omniSearch/animation/seeryCharacter.tsx
