+++ static/app/components/omniSearch/animation/useLottieCharacter.tsx
